fix: 감싼 링크를 표 분할 전에 개체에 고정한다 (#94)
`splitCell`은 문단과 개체를 **각자 다른 규칙**으로 조각에 배정한다 — 문단은
rect로 (걸치면 잘라서), 그림은 절단면에 걸치면 **양쪽에 복사**, 도형·글상자는
midY, 중첩 표는 minY. 그래서 U+FFFC run이 남지 않은 조각이 생기고, 그 조각의
개체는 (문단, 서수) 열쇠를 그대로 들고도 `wrapperHyperlinkURL`이 nil이 된다 —
방출과 히트가 같은 조회를 쓰므로 밑줄도 탭도 함께 사라졌다.

- 페이로드 4종에 `wrapperURL` 추가. **복사 헬퍼가 보존**한다 (R51 #2가 터졌던
  그 표면 — `withRect`/`withClip`/`withTable`/`offsetBy` 전부)
- `HwpTableCellFrame.resolvingWrapperURLs()` 가 쪼개기 **전에** 해석해 고정하고
  `splitCell`이 첫 줄에서 부른다. 이미 고정된 값은 덮지 않는다 (여러 페이지에
  걸친 표는 조각이 다시 쪼개진다)
- 히트·방출은 `wrapperURL ?? 키 조회` 순으로 본다

이 결함은 paint list 환원으로 안 풀린다 — 분할은 페이지네이션 중이고 paint
list는 그 뒤라, 짝이 깨진 뒤에는 어떤 소비자도 복원할 수 없다.

파일·타입 길이 한도(lint error)에 걸려 둘을 분리했다: `HwpCellContentFrames.swift`
(개체 페이로드 4종), `HwpTableWrapperLinkTests.swift` (감싼 링크 테스트 2종).
후자를 위해 `HwpTableLayoutTests`의 픽스처 도우미 익스텐션을 internal로 바꿨다.

검증: HwpKitCore 374개 0 실패. HWP_SNAPSHOT_TESTS=1로 렌더 해시 양 폰트 모드
무변화 + fidelity·골든·블록 스냅샷·페이지 수 무변화. 새 테스트는 `fillSegment`
로 실제 분할을 거치며, 분할 전 해석을 되돌리면 `wrapperURL`이 nil이 되고 히트가
`.table`로 떨어지는 것을 확인했다.
